requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
supabase==0.4.7
openai==1.6.1
